pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gtts>=2.5.4",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "sendgrid>=6.11.0",
    "elevenlabs>=1.56.0",
    "python-dotenv>=1.1.0",
    "qrcode>=8.1",
    "pillow>=11.1.0",
    "click>=8.1.8",
    "sqlalchemy>=2.0.40",
]
